extract code from editor mapitems into separate files
