+++ src/game/editor/editor.cpp
@@ -87,206 +87,6 @@ enum
 	BUTTON_CONTEXT = 1,
 };
 
-CEditorImage::~CEditorImage()
-{
-	m_pEditor->Graphics()->UnloadTexture(&m_Texture);
-	free(m_pData);
-	m_pData = nullptr;
-}
-
-CEditorSound::~CEditorSound()
-{
-	m_pEditor->Sound()->UnloadSample(m_SoundID);
-	free(m_pData);
-	m_pData = nullptr;
-}
-
-CLayerGroup::CLayerGroup()
-{
-	m_vpLayers.clear();
-	m_aName[0] = 0;
-	m_Visible = true;
-	m_Collapse = false;
-	m_GameGroup = false;
-	m_OffsetX = 0;
-	m_OffsetY = 0;
-	m_ParallaxX = 100;
-	m_ParallaxY = 100;
-	m_CustomParallaxZoom = 0;
-	m_ParallaxZoom = 100;
-
-	m_UseClipping = 0;
-	m_ClipX = 0;
-	m_ClipY = 0;
-	m_ClipW = 0;
-	m_ClipH = 0;
-}
-
-CLayerGroup::~CLayerGroup()
-{
-	m_vpLayers.clear();
-}
-
-void CLayerGroup::Convert(CUIRect *pRect)
-{
-	pRect->x += m_OffsetX;
-	pRect->y += m_OffsetY;
-}
-
-void CLayerGroup::Mapping(float *pPoints)
-{
-	float ParallaxZoom = m_pMap->m_pEditor->m_PreviewZoom ? m_ParallaxZoom : 100.0f;
-
-	m_pMap->m_pEditor->RenderTools()->MapScreenToWorld(
-		m_pMap->m_pEditor->MapView()->GetWorldOffset().x, m_pMap->m_pEditor->MapView()->GetWorldOffset().y,
-		m_ParallaxX, m_ParallaxY, ParallaxZoom, m_OffsetX, m_OffsetY,
-		m_pMap->m_pEditor->Graphics()->ScreenAspect(), m_pMap->m_pEditor->MapView()->GetWorldZoom(), pPoints);
-
-	pPoints[0] += m_pMap->m_pEditor->MapView()->GetEditorOffset().x;
-	pPoints[1] += m_pMap->m_pEditor->MapView()->GetEditorOffset().y;
-	pPoints[2] += m_pMap->m_pEditor->MapView()->GetEditorOffset().x;
-	pPoints[3] += m_pMap->m_pEditor->MapView()->GetEditorOffset().y;
-}
-
-void CLayerGroup::MapScreen()
-{
-	float aPoints[4];
-	Mapping(aPoints);
-	m_pMap->m_pEditor->Graphics()->MapScreen(aPoints[0], aPoints[1], aPoints[2], aPoints[3]);
-}
-
-void CLayerGroup::Render()
-{
-	MapScreen();
-	IGraphics *pGraphics = m_pMap->m_pEditor->Graphics();
-
-	if(m_UseClipping)
-	{
-		float aPoints[4];
-		m_pMap->m_pGameGroup->Mapping(aPoints);
-		float x0 = (m_ClipX - aPoints[0]) / (aPoints[2] - aPoints[0]);
-		float y0 = (m_ClipY - aPoints[1]) / (aPoints[3] - aPoints[1]);
-		float x1 = ((m_ClipX + m_ClipW) - aPoints[0]) / (aPoints[2] - aPoints[0]);
-		float y1 = ((m_ClipY + m_ClipH) - aPoints[1]) / (aPoints[3] - aPoints[1]);
-
-		pGraphics->ClipEnable((int)(x0 * pGraphics->ScreenWidth()), (int)(y0 * pGraphics->ScreenHeight()),
-			(int)((x1 - x0) * pGraphics->ScreenWidth()), (int)((y1 - y0) * pGraphics->ScreenHeight()));
-	}
-
-	for(auto &pLayer : m_vpLayers)
-	{
-		if(pLayer->m_Visible)
-		{
-			if(pLayer->m_Type == LAYERTYPE_TILES)
-			{
-				std::shared_ptr<CLayerTiles> pTiles = std::static_pointer_cast<CLayerTiles>(pLayer);
-				if(pTiles->m_Game || pTiles->m_Front || pTiles->m_Tele || pTiles->m_Speedup || pTiles->m_Tune || pTiles->m_Switch)
-					continue;
-			}
-			if(m_pMap->m_pEditor->m_ShowDetail || !(pLayer->m_Flags & LAYERFLAG_DETAIL))
-				pLayer->Render();
-		}
-	}
-
-	for(auto &pLayer : m_vpLayers)
-	{
-		if(pLayer->m_Visible && pLayer->m_Type == LAYERTYPE_TILES && pLayer != m_pMap->m_pGameLayer && pLayer != m_pMap->m_pFrontLayer && pLayer != m_pMap->m_pTeleLayer && pLayer != m_pMap->m_pSpeedupLayer && pLayer != m_pMap->m_pSwitchLayer && pLayer != m_pMap->m_pTuneLayer)
-		{
-			std::shared_ptr<CLayerTiles> pTiles = std::static_pointer_cast<CLayerTiles>(pLayer);
-			if(pTiles->m_Game || pTiles->m_Front || pTiles->m_Tele || pTiles->m_Speedup || pTiles->m_Tune || pTiles->m_Switch)
-			{
-				pLayer->Render();
-			}
-		}
-	}
-
-	if(m_UseClipping)
-		pGraphics->ClipDisable();
-}
-
-void CLayerGroup::AddLayer(const std::shared_ptr<CLayer> &pLayer)
-{
-	m_pMap->OnModify();
-	m_vpLayers.push_back(pLayer);
-}
-
-void CLayerGroup::DeleteLayer(int Index)
-{
-	if(Index < 0 || Index >= (int)m_vpLayers.size())
-		return;
-	m_vpLayers.erase(m_vpLayers.begin() + Index);
-	m_pMap->OnModify();
-}
-
-void CLayerGroup::DuplicateLayer(int Index)
-{
-	if(Index < 0 || Index >= (int)m_vpLayers.size())
-		return;
-
-	std::shared_ptr<CLayer> pDup = m_vpLayers[Index]->Duplicate();
-	m_vpLayers.insert(m_vpLayers.begin() + Index + 1, pDup);
-
-	m_pMap->OnModify();
-}
-
-void CLayerGroup::GetSize(float *pWidth, float *pHeight) const
-{
-	*pWidth = 0;
-	*pHeight = 0;
-	for(const auto &pLayer : m_vpLayers)
-	{
-		float lw, lh;
-		pLayer->GetSize(&lw, &lh);
-		*pWidth = maximum(*pWidth, lw);
-		*pHeight = maximum(*pHeight, lh);
-	}
-}
-
-int CLayerGroup::SwapLayers(int Index0, int Index1)
-{
-	if(Index0 < 0 || Index0 >= (int)m_vpLayers.size())
-		return Index0;
-	if(Index1 < 0 || Index1 >= (int)m_vpLayers.size())
-		return Index0;
-	if(Index0 == Index1)
-		return Index0;
-	m_pMap->OnModify();
-	std::swap(m_vpLayers[Index0], m_vpLayers[Index1]);
-	return Index1;
-}
-
-void CEditorImage::AnalyseTileFlags()
-{
-	mem_zero(m_aTileFlags, sizeof(m_aTileFlags));
-
-	int tw = m_Width / 16; // tilesizes
-	int th = m_Height / 16;
-	if(tw == th && m_Format == CImageInfo::FORMAT_RGBA)
-	{
-		unsigned char *pPixelData = (unsigned char *)m_pData;
-
-		int TileID = 0;
-		for(int ty = 0; ty < 16; ty++)
-			for(int tx = 0; tx < 16; tx++, TileID++)
-			{
-				bool Opaque = true;
-				for(int x = 0; x < tw; x++)
-					for(int y = 0; y < th; y++)
-					{
-						int p = (ty * tw + y) * m_Width + tx * tw + x;
-						if(pPixelData[p * 4 + 3] < 250)
-						{
-							Opaque = false;
-							break;
-						}
-					}
-
-				if(Opaque)
-					m_aTileFlags[TileID] |= TILEFLAG_OPAQUE;
-			}
-	}
-}
-
 void CEditor::EnvelopeEval(int TimeOffsetMillis, int Env, ColorRGBA &Channels, void *pUser)
 {
 	CEditor *pThis = (CEditor *)pUser;
@@ -7625,154 +7425,6 @@ void CEditor::Reset(bool CreateDefault)
 	m_SettingsCommandInput.Clear();
 }
 
-void CEditorMap::OnModify()
-{
-	m_Modified = true;
-	m_ModifiedAuto = true;
-	m_LastModifiedTime = m_pEditor->Client()->GlobalTime();
-}
-
-void CEditorMap::DeleteEnvelope(int Index)
-{
-	if(Index < 0 || Index >= (int)m_vpEnvelopes.size())
-		return;
-
-	OnModify();
-
-	VisitEnvelopeReferences([Index](int &ElementIndex) {
-		if(ElementIndex == Index)
-			ElementIndex = -1;
-		else if(ElementIndex > Index)
-			ElementIndex--;
-	});
-
-	m_vpEnvelopes.erase(m_vpEnvelopes.begin() + Index);
-}
-
-void CEditorMap::SwapEnvelopes(int Index0, int Index1)
-{
-	if(Index0 < 0 || Index0 >= (int)m_vpEnvelopes.size())
-		return;
-	if(Index1 < 0 || Index1 >= (int)m_vpEnvelopes.size())
-		return;
-	if(Index0 == Index1)
-		return;
-
-	OnModify();
-
-	VisitEnvelopeReferences([Index0, Index1](int &ElementIndex) {
-		if(ElementIndex == Index0)
-			ElementIndex = Index1;
-		else if(ElementIndex == Index1)
-			ElementIndex = Index0;
-	});
-
-	std::swap(m_vpEnvelopes[Index0], m_vpEnvelopes[Index1]);
-}
-
-template<typename F>
-void CEditorMap::VisitEnvelopeReferences(F &&Visitor)
-{
-	for(auto &pGroup : m_vpGroups)
-	{
-		for(auto &pLayer : pGroup->m_vpLayers)
-		{
-			if(pLayer->m_Type == LAYERTYPE_QUADS)
-			{
-				std::shared_ptr<CLayerQuads> pLayerQuads = std::static_pointer_cast<CLayerQuads>(pLayer);
-				for(auto &Quad : pLayerQuads->m_vQuads)
-				{
-					Visitor(Quad.m_PosEnv);
-					Visitor(Quad.m_ColorEnv);
-				}
-			}
-			else if(pLayer->m_Type == LAYERTYPE_TILES)
-			{
-				std::shared_ptr<CLayerTiles> pLayerTiles = std::static_pointer_cast<CLayerTiles>(pLayer);
-				Visitor(pLayerTiles->m_ColorEnv);
-			}
-			else if(pLayer->m_Type == LAYERTYPE_SOUNDS)
-			{
-				std::shared_ptr<CLayerSounds> pLayerSounds = std::static_pointer_cast<CLayerSounds>(pLayer);
-				for(auto &Source : pLayerSounds->m_vSources)
-				{
-					Visitor(Source.m_PosEnv);
-					Visitor(Source.m_SoundEnv);
-				}
-			}
-		}
-	}
-}
-
-void CEditorMap::MakeGameLayer(const std::shared_ptr<CLayer> &pLayer)
-{
-	m_pGameLayer = std::static_pointer_cast<CLayerGame>(pLayer);
-	m_pGameLayer->m_pEditor = m_pEditor;
-}
-
-void CEditorMap::MakeGameGroup(std::shared_ptr<CLayerGroup> pGroup)
-{
-	m_pGameGroup = std::move(pGroup);
-	m_pGameGroup->m_GameGroup = true;
-	str_copy(m_pGameGroup->m_aName, "Game");
-}
-
-void CEditorMap::Clean()
-{
-	m_vpGroups.clear();
-	m_vpEnvelopes.clear();
-	m_vpImages.clear();
-	m_vpSounds.clear();
-
-	m_MapInfo.Reset();
-	m_MapInfoTmp.Reset();
-
-	m_vSettings.clear();
-
-	m_pGameLayer = nullptr;
-	m_pGameGroup = nullptr;
-
-	m_Modified = false;
-	m_ModifiedAuto = false;
-
-	m_pTeleLayer = nullptr;
-	m_pSpeedupLayer = nullptr;
-	m_pFrontLayer = nullptr;
-	m_pSwitchLayer = nullptr;
-	m_pTuneLayer = nullptr;
-}
-
-void CEditorMap::CreateDefault(IGraphics::CTextureHandle EntitiesTexture)
-{
-	// add background
-	std::shared_ptr<CLayerGroup> pGroup = NewGroup();
-	pGroup->m_ParallaxX = 0;
-	pGroup->m_ParallaxY = 0;
-	pGroup->m_CustomParallaxZoom = 0;
-	pGroup->m_ParallaxZoom = 0;
-	std::shared_ptr<CLayerQuads> pLayer = std::make_shared<CLayerQuads>();
-	pLayer->m_pEditor = m_pEditor;
-	CQuad *pQuad = pLayer->NewQuad(0, 0, 1600, 1200);
-	pQuad->m_aColors[0].r = pQuad->m_aColors[1].r = 94;
-	pQuad->m_aColors[0].g = pQuad->m_aColors[1].g = 132;
-	pQuad->m_aColors[0].b = pQuad->m_aColors[1].b = 174;
-	pQuad->m_aColors[2].r = pQuad->m_aColors[3].r = 204;
-	pQuad->m_aColors[2].g = pQuad->m_aColors[3].g = 232;
-	pQuad->m_aColors[2].b = pQuad->m_aColors[3].b = 255;
-	pGroup->AddLayer(pLayer);
-
-	// add game layer and reset front, tele, speedup, tune and switch layer pointers
-	MakeGameGroup(NewGroup());
-	MakeGameLayer(std::make_shared<CLayerGame>(50, 50));
-	m_pGameGroup->AddLayer(m_pGameLayer);
-
-	m_pFrontLayer = nullptr;
-	m_pTeleLayer = nullptr;
-	m_pSpeedupLayer = nullptr;
-	m_pSwitchLayer = nullptr;
-	m_pTuneLayer = nullptr;
-}
-
 int CEditor::GetTextureUsageFlag()
 {
 	return Graphics()->HasTextureArrays() ? IGraphics::TEXLOAD_TO_2D_ARRAY_TEXTURE : IGraphics::TEXLOAD_TO_3D_TEXTURE;
@@ -8173,35 +7825,3 @@ void CEditor::LoadCurrentMap()
 }
 
 IEditor *CreateEditor() { return new CEditor; }
-
-// DDRace
-
-void CEditorMap::MakeTeleLayer(const std::shared_ptr<CLayer> &pLayer)
-{
-	m_pTeleLayer = std::static_pointer_cast<CLayerTele>(pLayer);
-	m_pTeleLayer->m_pEditor = m_pEditor;
-}
-
-void CEditorMap::MakeSpeedupLayer(const std::shared_ptr<CLayer> &pLayer)
-{
-	m_pSpeedupLayer = std::static_pointer_cast<CLayerSpeedup>(pLayer);
-	m_pSpeedupLayer->m_pEditor = m_pEditor;
-}
-
-void CEditorMap::MakeFrontLayer(const std::shared_ptr<CLayer> &pLayer)
-{
-	m_pFrontLayer = std::static_pointer_cast<CLayerFront>(pLayer);
-	m_pFrontLayer->m_pEditor = m_pEditor;
-}
-
-void CEditorMap::MakeSwitchLayer(const std::shared_ptr<CLayer> &pLayer)
-{
-	m_pSwitchLayer = std::static_pointer_cast<CLayerSwitch>(pLayer);
-	m_pSwitchLayer->m_pEditor = m_pEditor;
-}
-
-void CEditorMap::MakeTuneLayer(const std::shared_ptr<CLayer> &pLayer)
-{
-	m_pTuneLayer = std::static_pointer_cast<CLayerTune>(pLayer);
-	m_pTuneLayer->m_pEditor = m_pEditor;
-}
