+++ src/game/editor/mapitems/layer_game.cpp
@@ -1,6 +1,6 @@
 /* (c) Magnus Auvinen. See licence.txt in the root of the distribution for more information. */
 /* If you are missing that file, acquire a complete release at teeworlds.com.                */
-#include "editor.h"
+#include <game/editor/editor.h>
 
 CLayerGame::CLayerGame(int w, int h) :
 	CLayerTiles(w, h)
