+++ src/game/editor/mapitems/layer_quads.cpp
@@ -1,11 +1,6 @@
 /* (c) Magnus Auvinen. See licence.txt in the root of the distribution for more information. */
 /* If you are missing that file, acquire a complete release at teeworlds.com.                */
-#include <base/math.h>
-
-#include <engine/graphics.h>
-
-#include "editor.h"
-#include <game/client/render.h>
+#include <game/editor/editor.h>
 
 CLayerQuads::CLayerQuads()
 {
